requirementts.txt
# Task 02 - Image Generation Dependencies
# Stable Diffusion with PyTorch for M4 MacBook Pro

# Core PyTorch for Apple Silicon
torch>=2.0.0
torchvision>=0.15.0

# Diffusion models
diffusers>=0.21.0
accelerate>=0.20.0
transformers>=4.25.0

# Image processing
Pillow>=9.0.0

# Utilities
numpy>=1.21.0
safetensors>=0.3.0

# Note: This setup avoids JAX/Flax dependencies for better compatibility
# Optimized for CPU/MPS (Apple Silicon) usage